Introdused channel live/offline events (experimental)
